# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Team Project",
    "description": "This API allows you to create, read, update, and delete vehicles, stores, parts, and users.",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/vehicles/": {
      "get": {
        "tags": [
          "Vehicles"
        ],
        "summary": "Get all vehicles",
        "description": "Endpoint to get all vehicles",
        "parameters": [
          {
            "name": "vin",
            "in": "query",
            "description": "Vehicle Identification Number to filter vehicles by.",
            "required": false,
            "type": "string"
          },
          {
            "name": "make",
            "in": "query",
            "description": "Make of the vehicle to filter vehicles by.",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "Vehicles"
        ],
        "summary": "Create a vehicle",
        "description": "Endpoint to create a vehicle",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "make": {
                  "example": "any"
                },
                "model": {
                  "example": "any"
                },
                "year": {
                  "example": "any"
                },
                "color": {
                  "example": "any"
                },
                "price": {
                  "example": "any"
                },
                "condition": {
                  "example": "any"
                },
                "vin": {
                  "example": "any"
                },
                "base64Image": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "Vehicles"
        ],
        "summary": "Delete all vehicles",
        "description": "Endpoint to delete all vehicles",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  }
}